update modification extractor
