+++ dataset/vul4j_dataset.csv
@@ -88,7 +88,7 @@ no,vul_id,cve_id,cwe_id,cwe_name,owasp_id,repo_slug,human_patch,build_system,com
 87,VUL4J-87-S,,,,,apache/datasketches-java,https://github.com/apache/datasketches-java/commit/312a0882407cde6f241807bee6970341d581d2df,Maven,8,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,sketches,,,,,MS_EXPOSE_REP
 88,VUL4J-88-S,,,,,jenkinsci/zos-connector-plugin,https://github.com/jenkinsci/zos-connector-plugin/commit/1ecc64122b637dfb48c17846e7a3147ce0a5b9c6,Maven,7,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,root,,,,,MS_SHOULD_BE_FINAL
 89,VUL4J-89-S,,,,,semuxproject/semux-core,https://github.com/semuxproject/semux-core/commit/f2c79e8b5196c697850808873170a366b4086101,Maven,11,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,root,,,,,MS_MUTABLE_COLLECTION
-90,VUL4J-90-S,,,,,mcbrewster/test,https://github.com/mcbrewster/test/commit/e8dedb3de4e026b122357b3bbe26325271d03467,Maven,8,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,indexing-service,,,,,MS_MUTABLE_COLLECTION_PKGPROTECT
+90,VUL4J-90-S,,,,,mcbrewster/test,https://github.com/mcbrewster/test/commit/e8dedb3de4e026b122357b3bbe26325271d03467,Maven,8,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,indexing-service,,,,,MS_MUTABLE_COLLECTION_PKGPROTECT
 91,VUL4J-91-S,,,,,apache/commons-imaging,https://github.com/apache/commons-imaging/commit/27ca378cc06390e2a46c1de9df243dfd1c548015,Maven,7,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,root,,,,,MS_MUTABLE_ARRAY
 92,VUL4J-92-S,,,,,innovimax/xslthl,https://github.com/innovimax/xslthl/commit/ff0fd609e0f15c99897d8991e02a98d0f75272f1,Maven,8,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,root,,,,,MS_MUTABLE_COLLECTION_PKGPROTECT
 93,VUL4J-93-S,,,,,oVirt/ovirt-engine,https://github.com/oVirt/ovirt-engine/commit/5a0dd9b2bed1cee2a45e6638ef507ea5b03262c1,Maven,8,,mvn -DskipTests clean install,,,-Dhttps.protocols=TLSv1.2 -Denforcer.skip=true -Dcheckstyle.skip=true -Dcobertura.skip=true -DskipITs=true -Drat.skip=true -Dlicense.skip=true -Dpmd.skip=true -Dfindbugs.skip=true -Dgpg.skip=true -Dskip.npm=true -Dskip.gulp=true -Dskip.bower=true -Danimal.sniffer.skip=true -V -B,backend/manager/modules/compat,,,,,MS_MUTABLE_COLLECTION_PKGPROTECT
