+++ modification-extractor/src/main/java/org/vul4j/ExtractedData.java
@@ -1,0 +1,12 @@
+package org.vul4j;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class ExtractedData {
+    public List<String> methods = new ArrayList<>();
+    public List<String> attributes = new ArrayList<>();
+    public Map<String, ExtractedData> classes = new HashMap<>();
+}
